rename variable and files, thanks chatgpt
